+++ src/outlineview/alOutlineSortMode.ts
@@ -1,0 +1,5 @@
+export enum ALOutlineSortMode {
+    position = 0,
+    name = 1,
+    category = 3
+}