+++ src/symbollibraries/azSymbolAccessModifier.ts
@@ -1,0 +1,6 @@
+export enum AZSymbolAccessModifier {
+    Public = 0,
+    Local = 1,
+    Protected = 2,
+    Internal = 3    
+}