+++ src/symbolsinformation/methodInformation.ts
@@ -1,5 +1,7 @@
+import { AZSymbolAccessModifier } from "../symbollibraries/azSymbolAccessModifier";
+
 export class MethodInformation {
     name: string | undefined;
     header: string | undefined;
-    accessModifier: string | undefined;
+    accessModifier: AZSymbolAccessModifier | undefined;
 }