+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # Change Log
 
 ## 3.0.41
-
+ - Issue #6 - AL Outline panel - sort symbols by position, name and type
+ - Issue #327 - Use different icons in AL Outline panel
+ - Issue #328 - Improvements in AL Outline panel
+   - collapse all
+   - region support
+   - saving tree state
+   - more symbols collapsed by default
  - Issue #363 - XmlPorts - Generate headers 
    - new `Generate column headers for CSV export` action available on XmlPort header and XmlPort tableelement when XmlPort format is set to "VariableText"
  - Issue #377 - Don't add captions to the API pages fields
    - new `alOutline.createApiFieldsCaptions` setting added
  - Issue #378 - API Pages : API field name suggestions for 'number' fields (number instead of no)
+ - Issue #380 - AL Outline panel support Region Directive in AL (New Feature Request) 
  - Issue #385 - AL Language Server Crashing - temp-al-proxy.al 
    - all references to code creating that temp file removed
  - Issue #410 - Search in Object browser should not be case sensitive
@@ -30,8 +37,12 @@
  - Issue #425 - Variable Completion including double quotes
 
 Thank you
+ - GreatScott000 for reporting issue #6
+ - JavierFuentes for reporting issue #327 and #328
+ - NKarolak for issue #328 ideas
  - 4dimit for reporting issue #363
  - fvet for reporting issues #377, #378, #415 and #425
+ - louagej for reporting issue #380
  - pri-kise for reporting issue #385, #421
  - damse60 for reporting issue #410
  - guidorobben for reporting issue #412